tests: enable Github action
With this patch, three kinds of tests are triggered when there is a
merge request created in a Github kexec-tools repo,
   - format check using shfmt
   - static analysis using shellcheck
   - ShellSpec unit tests (test cases spec/)
   - integration tests ( tests cases in tests/)

The tests are run inside a docker image. This docker image has all the
needed software installed including fedpkg, shellspec and etc. This
Docker image also provides
/usr/share/cloud_images/Fedora-Cloud-Base-35-1.2.x86_64.qcow2 to avoid
repeatedly downloading the Fedora 35 cloud base image each time the
tests are triggered.